Extract getPreviewPostBySlug
