+++ examples/cms-datocms/pages/api/preview.js
@@ -1,4 +1,4 @@
-import fetchAPI from '../../lib/api'
+import { getPreviewPostBySlug } from '../../lib/api'
 
 export default async (req, res) => {
   // Check the secret and next parameters
@@ -11,24 +11,10 @@ export default async (req, res) => {
   }
 
   // Fetch the headless CMS to check if the provided `slug` exists
-  // getPostBySlug would implement the required fetching logic to the headless CMS
-  const data = await fetchAPI(
-    `
-  query PostBySlug($slug: String) {
-    post(filter: {slug: {eq: $slug}}) {
-      slug
-    }
-  }`,
-    {
-      preview: true,
-      variables: {
-        slug: req.query.slug,
-      },
-    }
-  )
+  const post = await getPreviewPostBySlug(req.query.slug)
 
   // If the slug doesn't exist prevent preview mode from being enabled
-  if (!data.post) {
+  if (!post) {
     return res.status(401).json({ message: 'Invalid slug' })
   }
 
@@ -37,6 +23,6 @@ export default async (req, res) => {
 
   // Redirect to the path from the fetched post
   // We don't redirect to req.query.slug as that might lead to open redirect vulnerabilities
-  res.writeHead(307, { Location: `/posts/${data.post.slug}` })
+  res.writeHead(307, { Location: `/posts/${post.slug}` })
   res.end()
 }
